+++ src/Helper.js
@@ -7,17 +7,17 @@ module.exports = {
 
   PLAYER_WITH_TV: ['Sonos Beam', 'Sonos Playbar', 'Sonos Playbase', 'Sonos Arc'],
 
-  REGEX_TIME: /([0-1][0-9]):([0-5][0-9]):([0-5][0-9])/, // Only hh:mm:ss and hours from 0 to 19
-  REGEX_TIME_DELTA: /^[-+]?([0-1][0-9]):([0-5][0-9]):([0-5][0-9])/, // Only +/- hh:mm:ss
+  REGEX_TIME: /^(([0-1][0-9]):([0-5][0-9]):([0-5][0-9]))$/, // Only hh:mm:ss and hours from 0 to 19
+  REGEX_TIME_DELTA: /^([-+]?([0-1][0-9]):([0-5][0-9]):([0-5][0-9]))$/, // Only +/- REGEX_TIME
   REGEX_IP: /^(?:(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9][0-9]|[0-9])(\.(?!$)|$)){4}$/,
   REGEX_SERIAL: /^([0-9a-fA-F][0-9a-fA-F]-){5}[0-9a-fA-F][0-9a-fA-F]:/, // the end might be improved
-  REGEX_RADIO_ID: /^[s][0-9]+$/,
+  REGEX_RADIO_ID: /^([s][0-9]+)$/,
   REGEX_2DIGITS: /^\d{1,2}$/, // up to 2 digits but at least 1
   REGEX_3DIGITS: /^\d{1,3}$/, // up to 3 digits but at least 1
   REGEX_2DIGITSSIGN: /^[-+]?\d{1,2}$/,
   REGEX_3DIGITSSIGN: /^[-+]?\d{1,3}$/,  
   REGEX_ANYCHAR: /.+/,  // any character but at least 1
-  REGEX_QUEUEMODES: /^(NORMAL|REPEAT_ONE|REPEAT_ALL|SHUFFLE|SHUFFLE_NOREPEAT|SHUFFLE_REPEAT_ONE)$/i,
+  REGEX_QUEUEMODES: /^(NORMAL|REPEAT_ONE|REPEAT_ALL|SHUFFLE|SHUFFLE_NOREPEAT|SHUFFLE_REPEAT_ONE)$/i,  // mixed case allowed
   REGEX_CSV: /^([a-zA-Z0-9äöüÄÖÜ]+)([ -]){0,1}([a-zA-Z0-9äöüÄÖÜ]+)(,([a-zA-ZäöüÄÖÜ0-9]+)([ -]){0,1}([a-zA-Z0-9äöüÄÖÜ]+))*$/,
 
   NRCSP_ERRORPREFIX: 'n-r-c-s-p: ',